update comment in course learning and course detail and progress in student dashboard

+++ spring-microservices/course-service/src/main/java/com/hoangphihiep/dto/response/EnrolledCoursesResponse.java
@@ -14,5 +14,5 @@ public class EnrolledCoursesResponse {
     private long classId;
     private String courseName;
     private String enrollmentDate;
-    private int progressPercentage;
+    private double progressPercentage;
 }
